refactor: extract app components to better modularization
Extracted all the apps components to a screens folder/module to better
modularization.
